Move all exceptions to thier own files and update their call accordingly.
